requirements.txt
streamlit~=1.15.1
joblib~=1.2.0
numpy~=1.21.5
Pillow~=8.4.0
scikit-learn==1.1.3